refactor: Make NullFrame and remove business logic in ResultView
